Make doc use the same request id as in doc + Remove unrelevant comment
